+++ core/ecschema-metadata/src/ecschema-metadata.ts
@@ -10,6 +10,7 @@ export * from "./Deserialization/SchemaGraphUtil";
 export * from "./Deserialization/JsonProps";
 export * from "./Deserialization/Helper";
 export * from "./Deserialization/XmlParser";
+export * from "./ECName";
 export * from "./ECObjects";
 export * from "./Editor/Editor";
 export * from "./Exception";
